CRiSp example OSM data for Bucharest

Data extracted from OpenStreetMap for examples used in the CRiSp package. All datasets are provided in a projected coordinate reference system (UTM 35), with exception for the bounding box, which is provided as latitude/ longitude coordinates (WGS84).

Usage

bucharest_osm

Format

A list of sf objects representing:

bb

The city bounding box.

boundary

The administrative boundary of Bucharest.

river_centerline

The Dâmbovița river centerline.

river_surface

The Dâmbovița river area.

aoi_network

The area of interest (AoI) to include the network around the river.

streets

The street network.

railways

The railway network.

aoi_buildings

The area of interest (AoI) to include buildings around the river.

buildings

The buildings in the corridor.

Source

https://www.openstreetmap.org/about