Cap corridor edges with a city boundary.

Cap corridor edges with a city boundary.

Usage

cap_corridor(corridor_edges, river, crs, bb, cap = "city")

Arguments

corridor_edges

Edge of the corridor as a simple feature

river

River centerline as a simple feature

crs

A coordinate reference system as an epsg code, e.g. 4326 for WGS84

bb

A bounding box as a matrix with 4 elements: xmin, ymin, xmax, ymax

cap

Character string with the type of cap to be used. Default is "city"

Value

A simple feature geometry