Delineate a corridor around a river.

Delineate a corridor around a river.

Usage

delineate(
  city_name,
  river_name,
  crs = NULL,
  network_buffer = NULL,
  buildings_buffer = NULL,
  dem_buffer = 2500,
  initial_method = "valley",
  initial_buffer = NULL,
  dem = NULL,
  max_iterations = 10,
  capping_method = "shortest-path",
  angle_threshold = 90,
  corridor = TRUE,
  segments = FALSE,
  riverspace = FALSE,
  force_download = FALSE,
  ...
)

Arguments

city_name

A place name as a string

river_name

A river name as a string

crs

The projected Coordinate Reference System (CRS) to use. If not provided, the suitable Universal Transverse Mercator (UTM) CRS is selected

network_buffer

Add a buffer (an integer in meters) around river to retrieve additional data (streets, railways, etc.). Default is 3000 m.

buildings_buffer

Add a buffer (an integer in meters) around the river to retrieve additional data (buildings). Default is 100 m.

dem_buffer

Add a buffer (an integer in meters) around the river to retrieve the DEM.

initial_method

The method employed to define the initial river corridor geometry. See initial_corridor() for the available methods

initial_buffer

Buffer region to add to the river geometry to setup the initial corridor (only used if initial_method is "buffer")

dem

Digital elevation model (DEM) of the region (only used if initial_method is "valley")

max_iterations

Maximum number of iterations employed to refine the corridor edges (see corridor_edge()).

capping_method

The method employed to connect the corridor edge end points (i.e. to "cap" the corridor). See cap_corridor() for the available methods

angle_threshold

Only network edges forming angles above this threshold (in degrees) are considered when forming segment edges. See delineate_segments() and rcoins::stroke(). Only used if segments is TRUE.

corridor

Whether to carry out the corridor delineation

segments

Whether to carry out the corridor segmentation

riverspace

Whether to carry out the riverspace delineation

force_download

Download data even if cached data is available

...

Additional (optional) input arguments for retrieving the DEM dataset (see get_dem()). Only relevant if initial_method is "valley" and dem is NULL

Value

A list with the corridor, segments, and riverspace geometries

Examples

delineate("Bucharest", "Dâmbovița")
#> The default `network_buffer` of 3000 m is used for corridor delineation.
#> Saving data to cache directory: /home/runner/.cache/CRiSp/osmdata_waterway__25.967_44.334_26.226_44.541.rds
#> Saving data to cache directory: /home/runner/.cache/CRiSp/osmdata_natural_water_25.967_44.334_26.226_44.541.rds
#> Saving data to cache directory: /home/runner/.cache/CRiSp/osmdata_highway_motorway_trunk_primary_secondary_tertiary_motorway_link_trunk_link_primary_link_secondary_link_tertiary_link_25.929_44.367_26.263_44.494.rds
#> Saving data to cache directory: /home/runner/.cache/CRiSp/osmdata_railway_rail_25.929_44.367_26.263_44.494.rds
#> Saving data to cache directory: /home/runner/.cache/CRiSp/dem_Copernicus_DSM_COG_10_N44_00_E026_00_DEM_Copernicus_DSM_COG_10_N44_00_E025_00_DEM_25.898_44.344_26.295_44.517.rds
#> Warning: to_spatial_subdivision assumes attributes are constant over geometries
#> $corridor
#> Geometry set for 1 feature 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 417802.4 ymin: 4914017 xmax: 439005.3 ymax: 4925645
#> Projected CRS: WGS 84 / UTM zone 35N
#> POLYGON ((438366.8 4915261, 438406 4915128, 438...
#> 
#> $segments
#> NULL
#> 
#> $riverspace
#> NULL
#>