Delineate a corridor around a river.

Delineate a corridor around a river.

Usage

delineate_corridor(
  city_name,
  river_name,
  crs = NULL,
  bbox_buffer = NULL,
  initial_method = "valley",
  initial_buffer = NULL,
  dem = NULL,
  capping_method = "direct",
  angle_threshold = 90,
  segments = FALSE,
  riverspace = FALSE,
  force_download = FALSE,
  ...
)

Arguments

city_name

A place name as a string

river_name

A river name as a string

crs

The projected Coordinate Reference System (CRS) to use. If not provided, the suitable Universal Transverse Mercator (UTM) CRS is selected

bbox_buffer

Add a buffer region to the city boundary to avoid edge effects close to its limits

initial_method

The method employed to define the initial river corridor geometry. See initial_corridor() for the available methods

initial_buffer

Buffer region to add to the river geometry to setup the initial corridor (only used if initial_method is "buffer")

dem

Digital elevation model (DEM) of the region (only used if initial_method is "valley")

capping_method

The method employed to connect the corridor edge end points (i.e. to "cap" the corridor). See cap_corridor() for the available methods

angle_threshold

Only network edges forming angles above this threshold (in degrees) are considered when forming segment edges. See get_segments() and rcoins::stroke(). Only used if segments is TRUE.

segments

Whether to carry out the corridor segmentation

riverspace

Whether to carry out the riverspace delineation

force_download

Download data even if cached data is available

...

Additional (optional) input arguments for retrieving the DEM dataset (see get_dem()). Only relevant if initial_method is "valley" and dem is NULL

Value

A simple feature geometry