Split a river corridor into segments

Segments are defined as corridor subregions separated by river-crossing transversal lines that form continuous strokes in the network.

Usage

delineate_segments(corridor, network, river_centerline, angle_threshold = 90)

Arguments

corridor

The river corridor as a simple feature geometry

network

The spatial network to be used for the segmentation

river_centerline

The river centerline as a simple feature geometry

angle_threshold

Only consider angles above this threshold (in degrees) to form continuous strokes in the network. See rcoins::stroke() for more details.

Value

Segment polygons as a simple feature geometry

Examples

corridor <- bucharest_delineation$corridor
network <- rbind(bucharest_osm$streets, bucharest_osm$railways) |>
  as_network()
#> Warning: to_spatial_subdivision assumes attributes are constant over geometries
river_centerline <- bucharest_osm$river_centerline
delineate_segments(corridor, network, river_centerline)
#> Geometry set for 10 features 
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 418298.9 ymin: 4914017 xmax: 438435.1 ymax: 4925645
#> Projected CRS: WGS 84 / UTM zone 35N
#> First 5 geometries:
#> POLYGON ((433409.6 4918083, 433427.9 4918079, 4...
#> POLYGON ((438011.7 4916483, 438036.3 4916400, 4...
#> POLYGON ((428079.6 4917639, 428075.1 4917641, 4...
#> POLYGON ((426539 4919514, 426525.9 4919502, 426...
#> POLYGON ((425140.9 4920580, 425130.9 4920580, 4...