Flatten a network by adding points at apparent intersections.

All crossing edges are identified, and the points of intersections are injected within the edge geometries. Note that the injected points are not converted to network nodes (this can be achieved via sfnetworks' sfnetworks::to_spatial_subdivision(), which is part of the tasks that are included in clean_network().

Usage

flatten_network(network)

Arguments

network

A network object

Value

A network object with additional points at intersections

Details

The functionality is similar to sfnetworks' sfnetworks::st_network_blend(), but in that case an external point is only injected to the closest edge.

Examples

edges <- dplyr::bind_rows(bucharest_osm$streets, bucharest_osm$railways)
network <- sfnetworks::as_sfnetwork(edges, directed = FALSE)
flatten_network(network)
#> # A sfnetwork with 4378 nodes and 4590 edges
#> #
#> # CRS:  EPSG:32635 
#> #
#> # An undirected multigraph with 459 components with spatially explicit edges
#> #
#> # Node data: 4,378 × 1 (active)
#>             geometry
#>          <POINT [m]>
#> 1 (432042.4 4918609)
#> 2 (432564.3 4918306)
#> 3 (425064.1 4918775)
#> 4   (424292 4918477)
#> 5 (425537.9 4917476)
#> 6 (426169.1 4918094)
#> # ℹ 4,372 more rows
#> #
#> # Edge data: 4,590 × 4
#>    from    to type                                                      geometry
#>   <int> <int> <chr>                                             <LINESTRING [m]>
#> 1     1     2 primary   (432042.4 4918609, 432051.4 4918605, 432065 4918598, 43…
#> 2     3     4 secondary (425064.1 4918775, 425039.1 4918763, 424987.3 4918742, …
#> 3     5     6 primary   (425537.9 4917476, 425549.1 4917487, 425639.3 4917573, …
#> # ℹ 4,587 more rows