Delineate a river corridor on a spatial network.

Delineate a river corridor on a spatial network.

Usage

get_corridor(
  network,
  river_centerline,
  river_surface,
  bbox,
  initial_method = "valley",
  buffer = NULL,
  dem = NULL,
  capping_method = "direct"
)

Arguments

network

The spatial network to be used for the delineation

river_centerline

A simple feature geometry representing the river centerline

river_surface

A simple feature geometry representing the river surface

bbox

Bounding box defining the extent of the area of interest

initial_method

The method employed to define the initial river corridor geometry. See initial_corridor() for the available methods

buffer

Buffer region to add to the river geometry to setup the initial corridor (only used if initial_method is "buffer")

dem

Digital elevation model (DEM) of the region (only used if initial_method is "valley")

capping_method

The method employed to connect the corridor edge end points (i.e. to "cap" the corridor). See cap_corridor() for the available methods

Value

A simple feature geometry representing the river corridor