Retrieve OpenStreetMap data for a given location

Retrieve OpenStreetMap data for a given location, including the city boundary, the river centreline and surface, the streets, the railways, and the buildings

Usage

get_osmdata(
  city_name,
  river_name,
  network_buffer = NULL,
  buildings_buffer = NULL,
  crs = NULL,
  force_download = FALSE
)

Arguments

city_name

A character string with the name of the city.

river_name

A character string with the name of the river.

network_buffer

Buffer distance in meters around the river to get the streets and railways, default is 0 means no network data will be downloaded

buildings_buffer

Buffer distance in meters around the river to get the buildings, default is 0 means no buildings data will be downloaded

crs

An integer with the EPSG code for the projection. If no CRS is specified, the default is the UTM zone for the city.

force_download

Download data even if cached data is available

Value

An list with the retrieved OpenStreetMap data sets for the given location

Examples

get_osmdata("Bucharest", "Dâmbovița")
#> Warning: Loading data from cache directory: /home/runner/.cache/CRiSp/osmdata_place_city_25.967_44.334_26.226_44.541.rds
#> Warning: Loading data from cache directory: /home/runner/.cache/CRiSp/osmdata_waterway_river_25.967_44.334_26.226_44.541.rds
#> Warning: Loading data from cache directory: /home/runner/.cache/CRiSp/osmdata_natural_water_25.967_44.334_26.226_44.541.rds
#> $bb
#>     xmin     ymin     xmax     ymax 
#> 25.96667 44.33425 26.22558 44.54140 
#> 
#> $boundary
#> Geometry set for 1 feature 
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 417769.7 ymin: 4909348 xmax: 438364 ymax: 4932424
#> Projected CRS: WGS 84 / UTM zone 35N
#> MULTIPOLYGON (((427347.8 4913289, 427386.6 4913...
#> 
#> $river_centerline
#> Geometry set for 1 feature 
#> Geometry type: MULTILINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 342590.9 ymin: 4897268 xmax: 457820 ymax: 5048321
#> Projected CRS: WGS 84 / UTM zone 35N
#> MULTILINESTRING ((342590.9 5043751, 342603 5043...
#> 
#> $river_surface
#> Geometry set for 1 feature 
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 418720.8 ymin: 4915950 xmax: 438326.9 ymax: 4924248
#> Projected CRS: WGS 84 / UTM zone 35N
#> MULTIPOLYGON (((427868.6 4920273, 427882.5 4920...
#>