Delineate a river corridor on a spatial network.

The corridor edges on the two river banks are drawn on the provided spatial network starting from an initial definition of the corridor (based e.g. on the river valley).

Usage

delineate_corridor(
  network,
  river_centerline,
  aoi = NULL,
  max_width = 2500,
  initial_method = "valley",
  buffer = NULL,
  dem = NULL,
  max_iterations = 10,
  capping_method = "shortest-path"
)

Arguments

network

The spatial network to be used for the delineation

river_centerline

A simple feature geometry representing the river centerline

aoi

Area of interest as sf object or bbox

max_width

(Approximate) maximum width of the regions considered on the two river banks

initial_method

The method employed to define the initial river corridor geometry. See initial_corridor() for the available methods

buffer

Buffer region to add to the river geometry to setup the initial corridor (only used if initial_method is "buffer")

dem

Digital elevation model (DEM) of the region (only used if initial_method is "valley")

max_iterations

Maximum number of iterations employed to refine the corridor edges (see corridor_edge()).

capping_method

The method employed to connect the corridor edge end points (i.e. to "cap" the corridor). See cap_corridor() for the available methods

Value

A simple feature geometry representing the river corridor

Examples

if (FALSE) { # interactive()
if (!requireNamespace("CRiSpData", quietly = TRUE)) {
  message("Install CRiSpData from GitHub to run this example.")
} else {
  network <- rbind(CRiSpData::bucharest_osm$streets,
                   CRiSpData::bucharest_osm$railways) |>
    as_network()
  crs <- get_utm_zone(CRiSpData::bucharest_osm$bb)
  aoi <- reproject(CRiSpData::bucharest_osm$bb, crs)
  delineate_corridor(network,
                     CRiSpData::bucharest_osm$river_centerline,
                     aoi,
                     dem = terra::unwrap(CRiSpData::bucharest_dem))
}
}